Changed README, added Dockerfile.local (again) and created scripts for setup and connecting to bash
README is now a LOT shorter, a LOT simpler, and should be much, much easier to use. Dockerfile.local is used in the build_local script, so PLEASE DO NOT TOUCH IT unless you know what you are doing. connect_local is used to simply connect to the containers without the hassle of checking if they are running or not.